user verification added user.topics component

+++ app/Livewire/User/TopicsComponent.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\User;
 
 use App\Models\Topic;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class TopicsComponent extends Component
@@ -20,20 +21,32 @@ class TopicsComponent extends Component
         $validated = $this->validate([
             'title' => 'required',
         ]);
+
+        //$this->topics = Topic::all()->where('user_id',Auth::user()->id); 
         $this->topic->title = $validated['title'];
-        $this->topic->save();
+        if ($this->topic->user_id == Auth::user()->id) {
+            $this->topic->save();
+        }
+        else redirect('logout');
+        
     }
 
     public function deleteTopic(): void{
         
         //$this->topic->();
         $id = $this->topic->id;
-        Topic::destroy($this->topic->id);
-        $this->dispatch("topicDelete", id: $id);
+        if ($this->topic->user_id == Auth::user()->id) {
+            Topic::destroy($this->topic->id);
+            $this->dispatch("topicDelete", id: $id);
+        }
+        
     }
 
     public function activateTopic(): void{        
-        $this->dispatch("activateTopic", id: $this->topic->id);
+        if ($this->topic->user_id == Auth::user()->id) {
+            $this->dispatch("activateTopic", id: $this->topic->id);
+        }
+        
     }
